requirements.txt
torch
torchvision
pillow
kaggle
